Move `FramRingArray.[hi]pp` to `FramSections/`
