+++ Sts1CobcSw/FramSections/FramRingArray.hpp
@@ -22,8 +22,6 @@
 #include <span>
 
 
-// TODO: Add cache in case the FRAM is not working
-// TODO: Move to FramSections
 namespace sts1cobcsw
 {
 template<typename T, Section ringArraySection, std::size_t nCachedElements>
