+++ Sts1CobcSw/FramSections/FramRingArray.ipp
