+++ iwyu.imp
@@ -47,13 +47,13 @@
   # Instead of all our .ipp files include the .hpp ones
   { include: ["\"Sts1CobcSw/Edu/Types.ipp\"",                           "private", "<Sts1CobcSw/Edu/Types.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FileSystem/FileSystem.ipp\"",               "private", "<Sts1CobcSw/FileSystem/FileSystem.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/FramSections/FramRingArray.ipp\"",          "private", "<Sts1CobcSw/FramSections/FramRingArray.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/PersistentVariables.ipp\"",    "private", "<Sts1CobcSw/FramSections/PersistentVariables.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/Subsections.ipp\"",            "private", "<Sts1CobcSw/FramSections/Subsections.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Hal/GpioPin.ipp\"",                         "private", "<Sts1CobcSw/Hal/GpioPin.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Hal/Spi.ipp\"",                             "private", "<Sts1CobcSw/Hal/Spi.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Hal/Uart.ipp\"",                            "private", "<Sts1CobcSw/Hal/Uart.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/Fram.ipp\"",                      "private", "<Sts1CobcSw/Periphery/Fram.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/Periphery/FramRingArray.ipp\"",             "private", "<Sts1CobcSw/Periphery/FramRingArray.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/ProgramId/ProgramId.ipp\"",                 "private", "<Sts1CobcSw/ProgramId/ProgramId.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Byte.ipp\"",                         "private", "<Sts1CobcSw/Serial/Byte.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Serial.ipp\"",                       "private", "<Sts1CobcSw/Serial/Serial.hpp>", "public"] },
@@ -73,6 +73,7 @@
   { include: ["\"Sts1CobcSw/FileSystem/LfsStorageDevice.hpp\"",         "public", "<Sts1CobcSw/FileSystem/LfsStorageDevice.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FileSystem/LfsWrapper.hpp\"",               "public", "<Sts1CobcSw/FileSystem/LfsWrapper.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/FramLayout.hpp\"",             "public", "<Sts1CobcSw/FramSections/FramLayout.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/FramSections/FramRingArray.hpp\"",          "public", "<Sts1CobcSw/FramSections/FramRingArray.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/PersistentVariableInfo.hpp\"", "public", "<Sts1CobcSw/FramSections/PersistentVariableInfo.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/PersistentVariables.hpp\"",    "public", "<Sts1CobcSw/FramSections/PersistentVariables.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/Section.hpp\"",                "public", "<Sts1CobcSw/FramSections/Section.hpp>", "public"] },
@@ -89,7 +90,6 @@
   { include: ["\"Sts1CobcSw/Periphery/Fram.hpp\"",                      "public", "<Sts1CobcSw/Periphery/Fram.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/FramEpsSpi.hpp\"",                "public", "<Sts1CobcSw/Periphery/FramEpsSpi.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/FramMock.hpp\"",                  "public", "<Sts1CobcSw/Periphery/FramMock.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/Periphery/FramRingArray.hpp\"",                  "public", "<Sts1CobcSw/Periphery/FramRingArray.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Periphery/Rf.hpp\"",                        "public", "<Sts1CobcSw/Periphery/Rf.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/ProgramId/ProgramId.hpp\"",                 "public", "<Sts1CobcSw/ProgramId/ProgramId.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Byte.hpp\"",                         "public", "<Sts1CobcSw/Serial/Byte.hpp>", "public"] },
